+++ packages/ripple-tide-api/src/mapping/tide-page-base-mapping.ts
@@ -1,4 +1,8 @@
 import {
+  map as topicTagsMapping,
+  includes as topicTagsIncludes
+} from './topic-tags/topic-tags-mapping.js'
+import {
   map as sidebarContactsMapping,
   includes as sidebarContactsIncludes
 } from './sidebar-contacts/sidebar-contacts-mapping.js'
@@ -16,11 +20,12 @@ import {
 } from './sidebar-site-section-nav/sidebar-site-section-nav-mapping.js'
 
 export const tidePageBaseMapping = ({
+  withTopicTags = false,
   withSidebarContacts = false,
   withSidebarRelatedLinks = false,
   withSidebarSocialShare = false,
   withSidebarSiteSectionNav = false
-}) => {
+} = {}) => {
   const sidebar: any = {}
 
   if (withSidebarContacts) {
@@ -42,20 +47,23 @@ export const tidePageBaseMapping = ({
   return {
     title: 'title',
     created: 'created',
-    modified: 'modified',
+    changed: 'changed',
     nid: 'id',
     sidebar: sidebar,
+    topicTags: topicTagsMapping,
     _src: (src) => (process.env.NODE_ENV === 'development' ? src : undefined)
   }
 }
 
 export const tidePageBaseIncludes = ({
+  withTopicTags = false,
   withSidebarContacts = false,
   withSidebarRelatedLinks = false,
   withSidebarSocialShare = false,
   withSidebarSiteSectionNav = false
-}) => {
+} = {}) => {
   return [
+    ...(withTopicTags ? topicTagsIncludes : []),
     ...(withSidebarContacts ? sidebarContactsIncludes : []),
     ...(withSidebarRelatedLinks ? sidebarRelatedLinksIncludes : []),
     ...(withSidebarSocialShare ? sidebarSocialShareIncludes : []),